requirements.txt
backcall==0.1.0
bcrypt==3.1.7
cffi==1.12.3
colorama==0.4.1
decorator==4.4.0
Django==1.10
django-extensions==2.2.1
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.4.2
python-dateutil==2.8.0
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
